refactor: extract component
