+++ front/src/views/Hasiera/HasieraContent.ts
@@ -1,0 +1,5 @@
+export interface HasieraContent {
+  deskribapena: string;
+  izenburua: string;
+  edukia: string;
+}
