+++ front/src/views/Hasiera/index.ts
@@ -1,0 +1,1 @@
+export { Hasiera } from './Hasiera';
